TRUNK GS: - some useless/duplicated code removed - changed some debug messages - removed some code warnings - changed visibility of some classes, methods and fields - some datatable class moved to datatables, according code style
git-svn-id: http://svn.l2jfree.com/l2jfree@712 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/L2_GameServer_c5/src/main/java/net/sf/l2j/gameserver/clientpackets/RequestBuyProcure.java
@@ -171,7 +171,7 @@ public class RequestBuyProcure extends ClientBasePacket {
     int nagrada=L2Manor.getInstance().getRewardItem(itemId,manor.getCastle().getCropReward(itemId));
 
     int cou= L2Manor.getInstance().getRewardAmount(itemId,manor.getCastle().getCropReward(itemId));
-    cou= (int) count / cou;
+    cou= count / cou;
             // Add item to Inventory and adjust update packet
             L2ItemInstance item;
             item = player.getInventory().addItem("Manor",nagrada , cou, player, manor);
